requirements.txt
openai-agents
openai
streamlit
uuid
pydantic
python-dotenv
asyncio
google-generativeai